+++ abstract_ranker/config.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+
+CACHE_DIR = Path("./.abstract_cache")
